+++ my-app/src/components/hashtag/style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const Container = styled.div`
+  padding: 4.8rem 0 6rem;
+`;
+
+export default Container;
